Move load class to ethereum_spec_tools
The load class can be used to load the json fixtures into spec objects. This can be re-used to build the evm tools like the t8n and b11r with some modifications. As such, it is beneficial to have this class in the ethereum_spec_tools package. This can then be imported into the testing framework. This commit moves the class from the testing framework into the ethereum spec tools and makes the necessary adjustments in terms of formatting and doc strings.
