+++ setup.cfg
@@ -13,6 +13,7 @@ classifiers =
 [options]
 packages =
     ethereum_spec_tools
+    ethereum_spec_tools/evm_tools
     ethereum_spec_tools/lint
     ethereum_spec_tools/lint/lints
     ethereum
