+++ src/ethereum_spec_tools/evm_tools/__init__.py
@@ -1,0 +1,3 @@
+"""
+Defines EVM tools for use in the Ethereum specification.
+"""
